backend/pyproject.toml
[project]
name = "gpt-dtskul-backend"
version = "0.1.0"
description = "Backend for gpt-dtskul"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "flitz>=0.0.7",
    "frontend>=0.0.3",
    "groq>=0.15.0",
    "jinja2==3.0.3",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pymupdf==1.25.2",
    "python-jose==3.3.0",
    "python-multipart==0.0.5",
    "pyyaml>=6.0.0",
    "sqlalchemy==2.0.34",
    "streamlit==1.41.1",
    "tools>=1.0.2",
    "uvicorn>=0.22.0",
]
